requirements.txt
pip==24.0
pyyaml==6.0.2
